Implementing Rails Internationalization (I18n)
The Ruby I18n (shorthand for internationalization) gem which is shipped with Ruby on Rails provides an easy-to-use and extensible framework for translating our application to a single custom language other than English or for providing multi-language support in your application.

The process of "internationalization" usually means to abstract all strings and other locale specific bits (such as date or currency formats) out of our application.